expose license hash
