+++ lib/licensee.rb
@@ -2,8 +2,13 @@ require_relative "licensee/version"
 require_relative "licensee/content_helper"
 require_relative "licensee/license"
 require_relative "licensee/project"
-require_relative "licensee/project_file"
 
+# Project files
+require_relative "licensee/project_file"
+require_relative "licensee/project_files/license_file.rb"
+require_relative "licensee/project_files/package_info.rb"
+
+# Matchers
 require_relative "licensee/matchers/exact_matcher"
 require_relative "licensee/matchers/copyright_matcher"
 require_relative "licensee/matchers/dice_matcher"
